+++ gbfs.cpp
@@ -1,7 +1,7 @@
 #include "gbfs.hpp"
 
 Solution solve_gbfs(vector<int> user_input, HeuristicCalculator h){
-    int puzzle_counter = 1;
+    unsigned long long puzzle_counter = 1;
     int num_expanded_nodes = 0;
     priority_queue<Puzzle*, vector<Puzzle*>, GBFSComparator>* open = new priority_queue<Puzzle*, vector<Puzzle*>, GBFSComparator>();
     unordered_set<unsigned long long>* closed = new unordered_set<unsigned long long>();
